delete all spot files
