--- backend/db/models/spot.js
@@ -1,67 +1,0 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Spot extends Model {
-
-    static associate(models) {
-      // Spot.belongsTo(models.User, {
-      //   foreignKey: 'ownerId'
-      // })
-    }
-    
-  }
-  Spot.init({
-    id: {
-      allowNull: false,
-      autoIncrement: true,
-      primaryKey: true,
-      type: DataTypes.INTEGER
-    },
-    ownerId: {
-      type: DataTypes.INTEGER,
-      allowNull: false
-    },
-    address: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    city: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    state: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    country: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    lat: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-    lng: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    description: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    price: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-  }, {
-    sequelize,
-    modelName: 'Spot',
-  });
-  return Spot;
-};
